% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_Gauss.R
\name{em.gauss.opti.groups}
\alias{em.gauss.opti.groups}
\title{Finding the optimal number of components}
\usage{
em.gauss.opti.groups(y, k, alpha, beta, method = "quantile",
  epsilon = 1e-06)
}
\arguments{
\item{y}{A data vector with observed values per bin.}

\item{k}{maximum numbers of components}

\item{alpha}{inverse gamma shape parameter}

\item{beta}{inverse gamma rate parameter}

\item{method}{method how startvalues should be evaluated. For more details see function CreateCluster}

\item{epsilon}{stopping criterion}
}
\value{
A list with  mu, var, pi, loglik, ecoff, AIC, BIC for each fitted distribution
}
\description{
This function provides the user information of the fitted distributions in order to choose the optimal number of components.
}
\details{
This function fits for each number of components (1:k) a mixing distribution of Gaussians by using
the function em.gauss. Furthermore, the fit of the distributions is measured by the two information criteria AIC and BIC.
}
\examples{
y <- c(2, 4, 5,6,5,2,2, 1, 1, 2,  2, 1,6,7,8,7,6, 5, 2,1)
em.gauss.opti.groups(y,
                     k= 2,
                     alpha = 1,
                     beta = 2)
}
