% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_Gauss.R
\name{em.gauss}
\alias{em.gauss}
\title{Estimating Ecoff by mixing Gaussians}
\usage{
em.gauss(y, mu, sigma2, pi, alpha, beta, epsilon = 1e-06,
  ecoff.quantile = 0.01, ecoff.comp = 0, max.iter = 1000)
}
\arguments{
\item{y}{A data vector with observed values per bin.}

\item{mu}{A vector with startvalues for mu}

\item{sigma2}{A vector with startvalues for the variance}

\item{pi}{A vector with startvalues for the mixing proportions}

\item{alpha}{Shape parameter of inverse gamma distribution}

\item{beta}{Scale parameter of inverse gamma distribution}

\item{epsilon}{Convergence criterium}

\item{ecoff.quantile}{Quantil which defines the ECOFF}

\item{ecoff.comp}{Component which is used for calculating the ECOFF}

\item{max.iter}{Maximum number of iterations}
}
\value{
A list with components mu, var, pi, loglik and ecoff
}
\description{
This function tries to find the Ecoff value of the input data, by aproximating the distribution
by an mixing distribution of Gaussian distributions
}
\details{
The data vector y is defined as one row of the EUCAST Data of the Zone Diameter Data.
The first value of y has to be the number of resistant observations.
Then the following elements are the number of observations in bin 6 to 6+length(y)-1.
A bin x includes all observations which had values form x-0.5 to x+0.5.

The function uses the EM Algorithm to fit a mixing distribution of Normals on the data.
Based on the result of the converged mixing distribution the algorithm evaluates the ECOFF value.
The ECOFF value is defined by default as the 0.01 quantile of the rightest distribution.
The rightest density is defined as the distribution, where the sum of the pis firstly is greater than 0.3.
Therefore, the distribution get ordered by their mean in ascending order and then the pis are commulated of the right.

Furthermore,  to avoid, that one of the sigma2 converges to 0, the likelihood get penalized in dependence of sigma2.
In detail, the penalization term follows a inverse gamma distribution with parameter alpha and beta.
}
\examples{
y <- c(2, 4, 5,6,5,2,2, 1, 1, 2,  2, 1,6,7,8,7,6, 5, 2,1)
em.gauss(y = y,
         mu = c(8.36, 17.28),
         sigma2 = c(1.67,  8.4),
         pi = c(1/2, 1/2),
         alpha = 1,
         beta = 3,
         epsilon = 0.0001)
}
