% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_Gauss.R
\name{plot_fct}
\alias{plot_fct}
\title{Plotting the fitted distribution with ECOFF}
\usage{
plot_fct(y, mu, sigma2, pi, ecoff)
}
\arguments{
\item{y}{A data vector with observed values per bin}

\item{mu}{A vector with values for mu}

\item{sigma2}{A vector with values for the variance}

\item{pi}{A vector with values for the mixing proportions}

\item{ecoff}{a numeric value of the ECOFF value}
}
\value{
a plot of the fitted distribution
}
\description{
This function plotts a histogram of the data and the fitted mixing distribubion of normals.
Furthermore, the Ecoff value is plotted by a line.
}
