% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZD.R
\docType{data}
\name{ZD}
\alias{ZD}
\title{Snapshot of Zone Data from EUCAST}
\format{A dataframe with 1476 rows and 51 columns}
\usage{
ZD
}
\description{
This data set a snapshot of the Zone Data from EUCAST from 30.5.2019
}
\details{
This data is a real dataset of the Zone Diameters from EUCAST.
The first two columns indicates the used Antibioticum and Bacterium.
Then the columns Z6 to Z50 indicates the number of  observations in that bin.
The column Z6 shows the number of resistent observations.
Furthermore, the datasets provides further information, like the ecoff value and the overall number of observations.
}
\keyword{datasets}
